Keep Codex subagent hook bridge private
