+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-cf29066e9465cb5ac1387d1d482d0939b9176220ecc69964da9af1a471939269  plugin-sdk-api-baseline.json
-ab43993cf713a96b191c55cf89bb215c18ecdc2d8edf50f31369ce3b162c56e3  plugin-sdk-api-baseline.jsonl
+e94933dd5638231fc6132e1a15e352d6c50736973119467a2a538033f5a48d2c  plugin-sdk-api-baseline.json
+499d8db081d97fbbed69bf606567efb0bf1a8e486ab390c76ec518dab748a299  plugin-sdk-api-baseline.jsonl
