+++ src/plugin-sdk/agent-harness-task-runtime.ts
@@ -14,19 +14,6 @@ import {
   resolveSubagentCompletionOrigin,
 } from "../agents/subagent-announce-delivery.js";
 import { resolveAnnounceOrigin } from "../agents/subagent-announce-origin.js";
-import {
-  SUBAGENT_ENDED_OUTCOME_ERROR,
-  SUBAGENT_ENDED_OUTCOME_KILLED,
-  SUBAGENT_ENDED_OUTCOME_OK,
-  SUBAGENT_ENDED_REASON_COMPLETE,
-  SUBAGENT_ENDED_REASON_ERROR,
-  SUBAGENT_ENDED_REASON_KILLED,
-  SUBAGENT_TARGET_KIND_SUBAGENT,
-  type SubagentLifecycleEndedOutcome,
-  type SubagentLifecycleEndedReason,
-} from "../agents/subagent-lifecycle-events.js";
-import { createSubsystemLogger } from "../logging/subsystem.js";
-import { getGlobalHookRunner } from "../plugins/hook-runner-global.js";
 import { normalizeOptionalString } from "../shared/string-coerce.js";
 import {
   assertAgentHarnessTaskRuntimeScope,
@@ -43,7 +30,6 @@ import { INTERNAL_MESSAGE_CHANNEL } from "../utils/message-channel.js";
 
 export type { TaskRecord as AgentHarnessTaskRecord };
 export type { AgentHarnessTaskRuntimeScope };
-export type { SubagentLifecycleEndedOutcome, SubagentLifecycleEndedReason };
 
 type AgentHarnessTaskRuntimeId = Parameters<typeof createRunningTaskRun>[0]["runtime"];
 type CreateRunningTaskRunParams = Parameters<typeof createRunningTaskRun>[0];
@@ -97,7 +83,6 @@ export type AgentHarnessCompletionDelivery = Awaited<
 >;
 
 const AGENT_HARNESS_COMPLETION_SOURCE_TOOL = "agent_harness_task";
-const log = createSubsystemLogger("plugin-sdk/agent-harness-task-runtime");
 
 export function createAgentHarnessTaskRuntime(
   params: AgentHarnessTaskRuntimeScopeParams,
@@ -233,130 +218,6 @@ export async function deliverAgentHarnessTaskCompletion(params: {
     directIdempotencyKey: buildAnnounceIdempotencyKey(params.announceId),
     signal: params.signal,
   });
-}
-
-export async function emitAgentHarnessSubagentSpawnedHook(params: {
-  scope: AgentHarnessTaskRuntimeScope;
-  runId: string;
-  childSessionKey: string;
-  agentId?: string;
-  label?: string;
-  mode: "run" | "session";
-  threadRequested: boolean;
-}): Promise<void> {
-  const hookRunner = getGlobalHookRunner();
-  if (!hookRunner?.hasHooks("subagent_spawned")) {
-    return;
-  }
-  try {
-    const scope = assertAgentHarnessTaskRuntimeScope(params.scope);
-    const requesterOrigin = resolveAgentHarnessRequesterOrigin(scope);
-    await hookRunner.runSubagentSpawned(
-      {
-        runId: params.runId,
-        childSessionKey: params.childSessionKey,
-        agentId: params.agentId?.trim() || "unknown",
-        label: normalizeOptionalString(params.label),
-        requester: {
-          channel: requesterOrigin?.channel,
-          accountId: requesterOrigin?.accountId,
-          to: requesterOrigin?.to,
-          threadId: requesterOrigin?.threadId,
-        },
-        threadRequested: params.threadRequested,
-        mode: params.mode,
-      },
-      {
-        runId: params.runId,
-        childSessionKey: params.childSessionKey,
-        requesterSessionKey: scope.requesterSessionKey,
-      },
-    );
-  } catch (error) {
-    log.warn("failed to emit agent harness subagent_spawned hook", {
-      runId: params.runId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-  }
-}
-
-export async function emitAgentHarnessSubagentEndedHook(params: {
-  scope: AgentHarnessTaskRuntimeScope;
-  runId: string;
-  targetSessionKey: string;
-  reason: SubagentLifecycleEndedReason;
-  outcome: SubagentLifecycleEndedOutcome;
-  endedAt?: number;
-  error?: string;
-}): Promise<void> {
-  const hookRunner = getGlobalHookRunner();
-  if (!hookRunner?.hasHooks("subagent_ended")) {
-    return;
-  }
-  try {
-    const scope = assertAgentHarnessTaskRuntimeScope(params.scope);
-    const requesterOrigin = resolveAgentHarnessRequesterOrigin(scope);
-    await hookRunner.runSubagentEnded(
-      {
-        targetSessionKey: params.targetSessionKey,
-        targetKind: SUBAGENT_TARGET_KIND_SUBAGENT,
-        reason: params.reason,
-        accountId: requesterOrigin?.accountId,
-        runId: params.runId,
-        endedAt: params.endedAt,
-        outcome: params.outcome,
-        error: params.error,
-      },
-      {
-        runId: params.runId,
-        childSessionKey: params.targetSessionKey,
-        requesterSessionKey: scope.requesterSessionKey,
-      },
-    );
-  } catch (error) {
-    log.warn("failed to emit agent harness subagent_ended hook", {
-      runId: params.runId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-  }
-}
-
-export function resolveAgentHarnessSucceededSubagentEnd(): {
-  reason: SubagentLifecycleEndedReason;
-  outcome: SubagentLifecycleEndedOutcome;
-} {
-  return {
-    reason: SUBAGENT_ENDED_REASON_COMPLETE,
-    outcome: SUBAGENT_ENDED_OUTCOME_OK,
-  };
-}
-
-export function resolveAgentHarnessFailedSubagentEnd(): {
-  reason: SubagentLifecycleEndedReason;
-  outcome: SubagentLifecycleEndedOutcome;
-} {
-  return {
-    reason: SUBAGENT_ENDED_REASON_ERROR,
-    outcome: SUBAGENT_ENDED_OUTCOME_ERROR,
-  };
-}
-
-export function resolveAgentHarnessKilledSubagentEnd(): {
-  reason: SubagentLifecycleEndedReason;
-  outcome: SubagentLifecycleEndedOutcome;
-} {
-  return {
-    reason: SUBAGENT_ENDED_REASON_KILLED,
-    outcome: SUBAGENT_ENDED_OUTCOME_KILLED,
-  };
-}
-
-function resolveAgentHarnessRequesterOrigin(scope: AgentHarnessTaskRuntimeScope) {
-  if (isInternalAnnounceRequesterSession(scope.requesterSessionKey)) {
-    return scope.requesterOrigin;
-  }
-  const { entry } = loadRequesterSessionEntry(scope.requesterSessionKey);
-  return resolveAnnounceOrigin(entry, scope.requesterOrigin);
 }
 
 function mapHarnessCompletionStatus(
